chore(PlanCalculator): move functions to plan.service.ts
