+++ src/i18n/en.js
@@ -324,7 +324,6 @@ You'll find an Email with steps to follow.`,
     discount_quarterly: 'Quaterly',
     discount_title: 'Subscription',
     discount_yearly: 'Yearly',
-    monthly_deliveries_amount_description: 'Emails',
     per_month: 'by month',
     plan_plus_title: 'Plan PLUS',
     plan_standard_title: 'Plan STANDARD',
@@ -334,13 +333,16 @@ You'll find an Email with steps to follow.`,
     plan_type_prepaid_tooltip: `Do you send Campaigns sporadically? Buy just the Credits you need for each sending. They won't expire and you can accumulate them!`,
     plan_type_subscribers: 'by Contacts',
     plan_type_subscribers_tooltip: 'Are you on a regular sending schedule? Choose a Plan based on the amount of Subscribers you have, with no sending limits. Get up to a 25% off, by paying 3, 6 or 12 months in advaced!',
-    prepaid_amount_description: 'Credits',
-    subscribers_amount_description: 'Contacts',
     subtitle: 'Choose a type plan and use the slider to calc the final price of your Plan.',
-    unknown_amount_description: ' ',
     with_half_yearly_discount: 'Half yearly payment will be of',
     with_quarterly_discount: 'Quarterly payment will be of',
     with_yearly_discount: 'Yearly payment will be of',
+  },
+  plans: {
+    monthly_deliveries_amount_description: 'Emails',
+    prepaid_amount_description: 'Créditos',
+    subscribers_amount_description: 'Contactos',
+    unknown_amount_description: ' ',
   },
   reports: {
     allow_enable_trial_MD: `Activate the trial period and access detailed Reports on the behavior of users inside your
