+++ src/i18n/es.js
@@ -323,7 +323,6 @@ Encontrarás un Email con los pasos a seguir.`,
     discount_quarterly: 'Trimestral',
     discount_title: 'Suscripción',
     discount_yearly: 'Anual',
-    monthly_deliveries_amount_description: 'Emails',
     per_month: 'por mes',
     plan_plus_title: 'Plan PLUS',
     plan_standard_title: 'Plan STANDARD',
@@ -333,13 +332,16 @@ Encontrarás un Email con los pasos a seguir.`,
     plan_type_prepaid_tooltip: '¿Envías Campañas en forma esporádica? Compra solo los Créditos que necesites para cada ocasión. ¡No tienen vencimiento y son acumulables!',
     plan_type_subscribers: 'por Contactos',
     plan_type_subscribers_tooltip: '¿Realizas envíos frecuentes? Escoge un Plan basado en la cantidad de Suscriptores que posees, sin límite de envíos. Si pagas 3, 6 o 12 meses por adelantado, ¡ahorras hasta un 25%!.',
-    prepaid_amount_description: 'Créditos',
-    subscribers_amount_description: 'Contactos',
     subtitle: 'Elige el tipo de plan y utiliza el slider para calcular el costo final de tu Plan.',
-    unknown_amount_description: ' ',
     with_half_yearly_discount: 'El pago semestral será de',
     with_quarterly_discount: 'El pago trimestral será de',
     with_yearly_discount: 'El pago anual será de',
+  },
+  plans: {
+    monthly_deliveries_amount_description: 'Emails',
+    prepaid_amount_description: 'Créditos',
+    subscribers_amount_description: 'Contactos',
+    unknown_amount_description: ' ',
   },
   reports: {
     allow_enable_trial_MD: `Activa el periodo de prueba y accede a Reportes detallados sobre el comportamiento de los
